use parGuess throughout bootstrap.metab
This function had been written such that the user passed in a parGuess vector of log-transformed parameter guesses as an argument, but also the function looked in the global environment for the log-transformed maximum likelihood parameter estimates as stored in optimTemp$par. This seems like it's creating problems with the function not using the right set of parameters in some applications, so trying to correct it. Now the user passes in the log-transformed ML estiamtes as an argument, and the function uses those estimates both to calculate the nominal predictions and residuals, and later as a starting point for subsequent runs of optim() on the bootstrapped data.